Remove author-/title-hints in doc
Remove the documentation for author and title hints for our script-based
processing from the `iacrdoc.tex`. Many authors use this file as a base
for writing their paper and do not delete the exemplary hints that set the
title and authors.

In fact, in the camera ready papers for TCHES v2020i2 we had 4
submissions that would have worked without hints, but were overwritten
by undeleted example hints.